i have no idea what I'm doing, I think I made a broken sh client and forked a broken ii

+++ clients/shidc/stdinbot.sh
@@ -1,9 +1,18 @@
 #!/bin/sh
 
-rm cow
-printf 'LOGIN\tUSERNAME=andrew\tPASSWORD=hunter2\r\n' > cow
+if [ -z "$4" ]
+then
+    printf '%s: Four arguments are required: Hostname, port, username and password\n' "$0" > /dev/stderr
+    exit 1
+fi
 
-(tail -n 1 -f cow | nc andrewyu.org 6835) &
+rm cow
+printf 'LOGIN\tUSERNAME=%s\tPASSWORD=%s\r\n' "$3" "$4" > cow
+
+(tail -n 1 -f cow | nc "$1" "$2") &
+ncid="$!"
+
+trap "kill $ncid" EXIT
 
 while read r
     do
@@ -11,3 +20,6 @@ while read r
     done
 
 fg
+
+# this little part doesn't work because sigint isn't caught
+kill "$ncid"
